broke out models.rb into seperate files
