+++ app/models/GameGenre.rb
@@ -1,0 +1,5 @@
+# A game genre has many games and many games have the same genre
+class GameGenre < ActiveRecord::Base
+    has_and_belongs_to_many :games
+    validates_presence_of :name
+end
