+++ app/models/VoteRecord.rb
@@ -1,0 +1,4 @@
+# A vote record belongs to a single game, which will have many of them
+class VoteRecord < ActiveRecord::Base
+    belongs_to :game
+end
